ClauseWise/requirements.txt
streamlit
transformers
spacy
pdfplumber
python-docx
docx2txt
PyPDF2

